requirements.txt
fastapi
uvicorn[standard]
python-dotenv

langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain-groq

chromadb
hnswlib

httpx
requests
numpy
os

